Remove context locks from visualization commands
It looks like visualization commands had required context lock
unnecessairly. Context manager methods are synchronized and therefore
should not need the lock before submitting a correspodning background
job.

Additionally, the presence of a context lock leads a deadlock:
1. Consider a long running execution of a program that does not finish
   within the 5 seconds
2. In the meantime there comes either an `AttachVisualization` or
   `DetachVisualization` request from the client

The latter will get stuck and eventually timeout out because it cannot
acquire the lock withing the required time limits.
It is still possible that a single long-running `ExecuteJob` might block
other ones (including visualization ones) but that's a separate work.

Fixes some issues present in #7941.
