+++ engine/runtime-instrument-common/src/main/scala/org/enso/interpreter/instrument/job/ExecuteJob.scala
@@ -86,7 +86,7 @@ class ExecuteJob(
       ctx.locking.releaseContextLock(contextId)
       logger.log(
         Level.FINEST,
-        s"Kept context lock [ExecuteJob] for ${contextId} for ${System.currentTimeMillis() - acquiredLock}"
+        s"Kept context lock [ExecuteJob] for ${contextId} for ${System.currentTimeMillis() - acquiredLock} milliseconds"
       )
 
     }
